WIP adding ability to handle PSD text layers
So far only most of the reading of the PSD data related to text handling
is done.
We still need to convert relevant parts to markup text that GIMP can
handle. To do that, we probably need to store all keys and values we
read, and then when converting to markup check all keys that we can
use for their values.